accflowbot/bot.py
"""AccFlow Shop — Telegram bot entry point."""

import asyncio
import logging
import os

from aiogram import Bot, Dispatcher
from dotenv import load_dotenv

from database import add_sample_products, init_db
from handlers import router

load_dotenv()

BOT_TOKEN = os.getenv("BOT_TOKEN")
if not BOT_TOKEN:
    raise SystemExit("Set BOT_TOKEN in .env")

logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
)
logger = logging.getLogger(__name__)


async def main() -> None:
    init_db()
    add_sample_products()

    bot = Bot(token=BOT_TOKEN)
    dp = Dispatcher()
    dp.include_router(router)

    logger.info("AccFlow Shop bot starting...")
    await dp.start_polling(bot)


if __name__ == "__main__":
    asyncio.run(main())
